refactor: Update default values for username and hostname in new.py
The code changes update the default values for the `username` and `hostname` variables in `new.py`. The `username` now defaults to "unho.chang" and the `hostname` defaults to "dataotaku.com". This change improves the default behavior of the program by using more meaningful values for the username and hostname when environment variables are not set.